--- product_agent_streamlit/adapters/__init__.py
@@ -1,7 +1,0 @@
-"""
-Adapters package for integrating different AI models with OpenAI Agents SDK
-"""
-
-from .gemini_model import GeminiChatModel
-
-__all__ = ['GeminiChatModel']
