+++ product_agent_streamlit/custom_agents/__init__.py
@@ -1,8 +1,4 @@
-"""
-Agents package for the Multi-Catalog System
-"""
-
+from .catalog_agent import CatalogAgent
 from .orchestrator_agent import OrchestratorAgent
-from .catalog_agent import PDFCatalogAgent
 
-__all__ = ['OrchestratorAgent', 'PDFCatalogAgent']
+__all__ = ["CatalogAgent", "OrchestratorAgent"]
