--- product_agent_streamlit/models/__init__.py
@@ -1,12 +1,0 @@
-# # File: models/__init__.py
-# from .catalog_models import PDFMetadata, CatalogSearchResult, ChatMessage
-
-# __all__ = ['PDFMetadata', 'CatalogSearchResult', 'ChatMessage']
-
-"""
-Models package for the Multi-Catalog System
-"""
-
-from .pdf_metadata import PDFMetadata
-
-__all__ = ['PDFMetadata']
