+++ product_agent_streamlit/processors/__init__.py
@@ -2,6 +2,7 @@
 Processors package for the Multi-Catalog System
 """
 
-from .pdf_processor import PDFCatalogProcessor
+from .pdf_processor import PDFProcessor
+from .catalog_manager import CatalogManager
 
-__all__ = ['PDFCatalogProcessor']
+__all__ = ['PDFProcessor', 'CatalogManager']
