--- product_agent_streamlit/storage/__init__.py
@@ -1,7 +1,0 @@
-"""
-Storage package for the Multi-Catalog System
-"""
-
-from .catalog_library import CatalogLibrary
-
-__all__ = ['CatalogLibrary']
