--- product_agent_streamlit/tools/__init__.py
@@ -1,8 +1,0 @@
-"""
-Tools package for the Multi-Catalog System
-"""
-
-from .orchestrator_tools import OrchestratorTools
-from .catalog_tools import CatalogTools
-
-__all__ = ['OrchestratorTools', 'CatalogTools']
