--- product_agent_streamlit/ui/__init__.py
@@ -1,4 +1,0 @@
-# File: ui/__init__.py
-from .streamlit_app import create_streamlit_app
-
-__all__ = ['create_streamlit_app']
